HTTP response
The current entry for "HTTP response" starts by describing the request process, which definition exists right above.

A description of "response" which starts with, and emphasizes, the response concept itself is proposed.

Moreover, the explanation implies that the content of the body of a webpage is the only expected body here in the case of a website. However, websites serve all types of content in HTTP responses, including images and all other sorts of files. So perhaps this is also due to be noted to avoid ambiguity.
That is without taking into consideration cases where HTTP is used for other purposes, which is suitably ignored here, for the sake of simplification, and which is assumed the reason the author specified "the case of website".

On a side note, perhaps a cross-reference is also a good idea here (as per #1348) . Alternatively, a comprehensive description of the request/response process and concepts could be placed in the first occurring term, i.e "request", and a "see the definition of 'request'", preferably with a link, can be put in the other.